+++ lib/screens/sign_in_screen.dart
@@ -12,8 +12,8 @@ import 'package:sul_sul/models/auth_repository.dart';
 
 import 'package:sul_sul/utils/api/http_code.dart';
 
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({super.key});
+class SignInScreen extends StatelessWidget {
+  const SignInScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
